[Wrench Est.][Plugin]: separate i_term as an independent file
